third_party/WebKit/LayoutTests/shadow-dom/v0/event-path.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<img id="img" src="../../images/resources/test-load.jpg">
<script>

// Basic functionality test for Event.path.
img.onload = function(e) {
    test(function() {
        assert_array_equals(e.composedPath(), e.path);
    }, "Event.composedPath() should be the same as Event.path.");
};

</script>
